Work in progress on setting up parallel merging

+++ merge.hpp
@@ -7,4 +7,4 @@ struct merge {
 	bool operator<(const merge &m) const;
 };
 
-void do_merge(struct merge *merges, int nmerges, FILE *f, int bytes, long long nrec, int zoom);
+void do_merge(struct merge *merges, size_t nmerges, FILE *f, int bytes, long long nrec, int zoom);
